IS: Refactor oppfolgingsoppgave into personstatus folder
