+++ src/main/kotlin/no/nav/syfo/personstatus/application/oppfolgingsoppgave/OppfolgingsoppgaveService.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo.oppfolgingsoppgave
+package no.nav.syfo.personstatus.application.oppfolgingsoppgave
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -7,36 +7,45 @@ import no.nav.syfo.personstatus.infrastructure.database.DatabaseInterface
 import no.nav.syfo.personstatus.application.PersonBehandlendeEnhetService
 import no.nav.syfo.personstatus.domain.PersonIdent
 import no.nav.syfo.personstatus.infrastructure.database.queries.createPersonOversiktStatus
-import no.nav.syfo.oppfolgingsoppgave.domain.Oppfolgingsoppgave
-import no.nav.syfo.oppfolgingsoppgave.kafka.COUNT_KAFKA_CONSUMER_TRENGER_OPPFOLGING_READ
+import no.nav.syfo.personstatus.domain.PersonOversiktStatus
+import no.nav.syfo.personstatus.infrastructure.kafka.oppfolgingsoppgave.COUNT_KAFKA_CONSUMER_TRENGER_OPPFOLGING_READ
 import no.nav.syfo.personstatus.infrastructure.database.queries.getPersonOversiktStatusList
 import no.nav.syfo.personstatus.infrastructure.database.queries.updateOppfolgingsoppgave
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
+import kotlin.collections.filter
+import kotlin.collections.first
+import kotlin.collections.firstOrNull
+import kotlin.collections.forEach
+import kotlin.jvm.java
+import kotlin.use
 
 class OppfolgingsoppgaveService(
     private val database: DatabaseInterface,
     private val personBehandlendeEnhetService: PersonBehandlendeEnhetService,
 ) {
-    fun processOppfolgingsoppgave(records: List<Oppfolgingsoppgave>) {
+    fun processOppfolgingsoppgave(records: List<OppfolgingsoppgaveRecord>) {
         createOrUpdatePersonOversiktStatus(records)
         records.filter { it.isActive }.forEach {
             val personOversiktStatus = database.getPersonOversiktStatusList(
-                fnr = it.personIdent.value,
+                fnr = it.personIdent,
             ).first()
-            updateBehandlendeEnhet(it.personIdent, personOversiktStatus.enhet)
+            updateBehandlendeEnhet(PersonIdent(it.personIdent), personOversiktStatus.enhet)
         }
     }
 
-    private fun createOrUpdatePersonOversiktStatus(records: List<Oppfolgingsoppgave>) {
+    private fun createOrUpdatePersonOversiktStatus(records: List<OppfolgingsoppgaveRecord>) {
         database.connection.use { connection ->
             records.forEach { oppfolgingsOppgave ->
                 val existingPersonOversiktStatus = connection.getPersonOversiktStatusList(
-                    fnr = oppfolgingsOppgave.personIdent.value,
+                    fnr = oppfolgingsOppgave.personIdent,
                 ).firstOrNull()
 
                 if (existingPersonOversiktStatus == null) {
-                    val personoversiktStatus = oppfolgingsOppgave.toPersonoversiktStatus()
+                    val personoversiktStatus = PersonOversiktStatus(
+                        fnr = oppfolgingsOppgave.personIdent,
+                        isAktivOppfolgingsoppgave = oppfolgingsOppgave.isActive,
+                    )
                     connection.createPersonOversiktStatus(
                         commit = false,
                         personOversiktStatus = personoversiktStatus,
