+++ src/main/kotlin/no/nav/syfo/personstatus/infrastructure/kafka/KafkaModule.kt
@@ -16,7 +16,7 @@ import no.nav.syfo.personstatus.application.PersonoversiktStatusService
 import no.nav.syfo.personstatus.application.OppfolgingstilfelleService
 import no.nav.syfo.personstatus.infrastructure.kafka.manglendemedvirkning.ManglendeMedvirkningVurderingConsumer
 import no.nav.syfo.personstatus.infrastructure.kafka.meroppfolging.SenOppfolgingKandidatStatusConsumer
-import no.nav.syfo.oppfolgingsoppgave.kafka.launchOppfolgingsoppgaveConsumer
+import no.nav.syfo.personstatus.infrastructure.kafka.oppfolgingsoppgave.launchOppfolgingsoppgaveConsumer
 
 fun launchKafkaModule(
     applicationState: ApplicationState,
